requirements.txt
arrow==1.3.0
asgiref==3.5.2
canvasapi==3.3.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
colorama==0.4.5
Django==4.0.6
django-modeltranslation==0.18.4
django-registration-redux==2.11
et_xmlfile==2.0.0
idna==3.10
openpyxl==3.1.5
Pillow==9.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
qrcode==7.3.1
reportlab==4.3.1
requests==2.32.3
six==1.17.0
sqlparse==0.4.2
types-python-dateutil==2.9.0.20241206
tzdata==2022.1
urllib3==2.4.0
xlwt==1.3.0
